requirements.txt
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
Pillow==10.2.0
chardet==5.2.0
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai>=0.8.3
python-dotenv==1.0.0
